+++ src/Decode.h
@@ -9,7 +9,7 @@
 namespace Decode {
     uint16_t to_uint16_t (const char *buffer);
     uint32_t to_uint32_t (const char *buffer);
-    int read_varint (std::ifstream *db, uint16_t starting_offset, uint16_t &new_offset);
+    uint64_t read_varint (std::ifstream *db, size_t starting_offset, size_t &new_offset);
 }
  
 #endif
